pyproject.toml
[tool.poetry]
name = "tabletop-board-backend"
version = "0.1.0"
description = "Service to find players for offline tabletop games"
authors = ["Sergei Konik <s.konik.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
isort = "^5.8.0"
safety = "^1.10.3"
pytest-django = "^4.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
mypy = "^0.812"
django-stubs = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "django"
src_paths = ["backend"]